% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfa.R
\docType{data}
\name{nfa}
\alias{nfa}
\title{DATASET_TITLE}
\format{
A data frame with 4743 rows and 33 variables:
\describe{
\item{\code{country}}{character COLUMN_DESCRIPTION}
\item{\code{ifs_code}}{International Financial Statistics Country Codes}
\item{\code{date}}{Year}
\item{\code{ipeqa}}{double Portfolio equity assets (Stock)}
\item{\code{ipeql}}{double Portfolio equity liabilities (Stock)}
\item{\code{ipfdia}}{double Foreign Direct Investment Assets (Stock)}
\item{\code{ipfdil}}{double Foreign Direct Investment Liabilities}
\item{\code{ipdebta}}{double Debt assets (Stock)}
\item{\code{ipdebtl}}{double Debt liabilities (Stock)}
\item{\code{ipdera}}{double Financial Derivatives (Assets)}
\item{\code{ipderl}}{double Financial Derivatives (Liabilities)}
\item{\code{ipfx}}{double Non-gold Foreign Exchange Reserves}
\item{\code{ipfa}}{double Total assets}
\item{\code{ipfl}}{double Total liabilities}
\item{\code{ipnfa}}{double Non-gold Net International Investment Position}
\item{\code{iip}}{double Official international investment position}
\item{\code{nfa_alt}}{double Non-gold Net International Investment Position with an alternative FDI or equity valuation}
\item{\code{gdpd}}{double GDP (US$)}
\item{\code{nfa2y}}{double Ratio of Non-gold net international investment position to GDP}
\item{\code{ca_weo}}{double Current Account Balance (World Economic Outlook)}
\item{\code{ca}}{double Current account balance}
\item{\code{ka}}{double Capital account}
\item{\code{ippda}}{double Portfolio debt assets}
\item{\code{ippdl}}{double Portfolio debt liabilities}
\item{\code{ipoa}}{double Other investment assets}
\item{\code{ipol}}{double Other investment liabilities}
\item{\code{ipfdia_oth}}{double Other Foreign Direct Investment Assets}
\item{\code{ipfdil_oth}}{double Other Foreign Direct Investment Liabilities}
\item{\code{e_avg}}{double Average exchange rate during period}
\item{\code{e_eop}}{double Exchange rate at the end of period}
\item{\code{nfa2y_lc}}{double Ratio of Net non-gold international investment position to GDP (Domestic currency)}
\item{\code{iip2y_lc}}{double Ratio of net international investment position to GDP (Domestic Currency)}
\item{\code{source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
nfa
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
